# services/main/docs/swagger.yaml
definitions:
...
  lib.HTTPError:
    properties:
      errors:
        properties:
          message:
            type: string
        type: object
    type: object
  lib.HTTPReturnPaginatedMetaResponse:
    properties:
      has_next_page:
        example: true
        type: boolean
      has_previous_page:
        example: false
        type: boolean
      order:
        example: desc
        type: string
      order_by:
        example: created_at
        type: string
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
info:
  contact:
    email: domeybenjamin1@gmail.com
    name: Domey Benjamin
    url: https://www.linkedin.com/in/benjamin-armah-kesey-domey/
  description: This is the RentLoop API documentation.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://rentloop.com/terms
  title: RentLoop API
  version: "1.0"
paths:
...
  /api/v1/admin/properties/{property_id}/blocks/{block_id}/units:
    post:
      consumes:
      - application/json
      description: Create unit (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Property block ID
        in: path
        name: block_id
        required: true
        type: string
      - description: Unit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateUnitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.AdminOutputUnit'
            type: object
        "400":
          description: Error occurred when creating a unit
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "422":
          description: Invalid request body
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create unit (Admin)
      tags:
      - Units
...
  /api/v1/admin/properties/{property_id}/units:
    get:
      consumes:
      - application/json
      description: List units (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - collectionFormat: multi
        example:
        - 767d8e23-8c9f-4c51-85af-5908039869da
        - 3d90d606-2a22-4487-9431-69736829094f
        in: query
        items:
          type: string
        name: block_ids
        type: array
      - in: query
        name: end_date
        type: string
      - collectionFormat: multi
        example:
        - a8098c1a-f86e-11da-bd1a-00112444be1e
        in: query
        items:
          type: string
        name: ids
        type: array
      - enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      - in: query
        name: order_by
        type: string
      - in: query
        minimum: 0
        name: page
        type: integer
      - in: query
        minimum: 0
        name: page_size
        type: integer
      - enum:
        - WEEKLY
        - DAILY
        - MONTHLY
        - QUARTERLY
        - BIANNUALLY
        - ANNUALLY
        example: WEEKLY
        in: query
        name: payment_frequency
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: populate
        type: array
      - example: prop_123
        in: query
        name: property_id
        type: string
      - in: query
        name: query
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        minItems: 1
        name: search_fields
        type: array
      - in: query
        name: start_date
        type: string
      - enum:
        - Unit.Status.Draft
        - Unit.Status.Available
        - Unit.Status.Occupied
        - Unit.Status.Maintenance
        example: Unit.Status.Available
        in: query
        name: status
        type: string
      - enum:
        - APARTMENT
        - HOUSE
        - STUDIO
        - OFFICE
        - RETAIL
        example: SINGLE
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                properties:
                  meta:
                    $ref: '#/definitions/lib.HTTPReturnPaginatedMetaResponse'
                  rows:
                    items:
                      $ref: '#/definitions/transformations.AdminOutputUnit'
                    type: array
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: List units (Admin)
      tags:
      - Units
  /api/v1/admin/properties/{property_id}/units/{unit_id}:
    delete:
      consumes:
      - application/json
      description: delete a unit (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Unit ID
        in: path
        name: unit_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Unit deleted successfully
        "400":
          description: Error occurred when deleting a unit
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: Unit not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: delete a unit (Admin)
      tags:
      - Units
    get:
      consumes:
      - application/json
      description: Get unit (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Unit ID
        in: path
        name: unit_id
        required: true
        type: string
      - collectionFormat: csv
        in: query
        items:
          type: string
        name: populate
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: Unit retrieved successfully
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.AdminOutputUnit'
            type: object
        "400":
          description: Error occurred when fetching a unit
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "404":
          description: Unit not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get unit (Admin)
      tags:
      - Units
    patch:
      consumes:
      - application/json
      description: Update unit (Admin)
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Unit ID
        in: path
        name: unit_id
        required: true
        type: string
      - description: Unit
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateUnitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.AdminOutputUnit'
            type: object
        "400":
          description: Error occurred when updating a unit
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: Unit not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "422":
          description: Invalid request body
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update unit (Admin)
      tags:
      - Units
  /api/v1/admin/properties/{property_id}/units/{unit_id}/status:available:
    patch:
      consumes:
      - application/json
      description: Update unit (Admin) status to available
      parameters:
      - description: Property ID
        in: path
        name: property_id
        required: true
        type: string
      - description: Unit ID
        in: path
        name: unit_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Unit status updated successfully
        "400":
          description: Error occurred when updating unit status
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "401":
          description: Invalid or absent authentication token
          schema:
            type: string
        "403":
          description: Forbidden access
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: Unit not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update unit (Admin) status to available
      tags:
      - Units
...
  /api/v1/units/{unit_id}:
    get:
      consumes:
      - application/json
      description: Fetch unit subset
      parameters:
      - description: Unit ID
        in: path
        name: unit_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              data:
                $ref: '#/definitions/transformations.OutputUnit'
            type: object
        "400":
          description: Error occurred when fetching a unit
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "404":
          description: Unit not found
          schema:
            $ref: '#/definitions/lib.HTTPError'
        "500":
          description: An unexpected error occurred
          schema:
            type: string
      summary: Fetch unit subset
      tags:
      - Units
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"